refactor(playlist overview): create folder to encapsulate all components
